proto: Fix typo in message name
RecoderChannel::data message is the same as PlaybackChannel::data
one.
This fixed compilation of spice-gtk using updated spice-common.

Before generating the messages automatically the definition was:

  typedef struct SpiceMsgPlaybackPacket {
      uint32_t time;
      uint8_t *data;
      uint32_t data_size;
  } SpiceMsgPlaybackPacket, SpiceMsgcRecordPacket;

so both messages were defined as struct SpiceMsgPlaybackPacket with
SpiceMsgcRecordPacket an alias to it.

But spice.proto for RecordChannel::data was changed from
SpiceMsgcRecordPacket (the alias) to SpiceMsgcPlaybackPacket
defining a new structure and breaking the alias.

Signed-off-by: Frediano Ziglio <fziglio@redhat.com>
Acked-by: Victor Toso <victortoso@redhat.com>